[release test] remove JobRunner
merge everything into AnyscaleJobRunner
