+++ release/ray_release/command_runner/anyscale_job_runner.py
@@ -2,6 +2,7 @@ import json
 import os
 import re
 import shlex
+import shutil
 import tempfile
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
@@ -11,7 +12,7 @@ from ray_release.cloud_util import (
     upload_working_dir_to_azure,
 )
 from ray_release.cluster_manager.cluster_manager import ClusterManager
-from ray_release.command_runner.job_runner import JobRunner
+from ray_release.command_runner.command_runner import CommandRunner
 from ray_release.exception import (
     FetchResultError,
     JobBrokenError,
@@ -19,6 +20,7 @@ from ray_release.exception import (
     JobOutOfRetriesError,
     JobTerminatedBeforeStartError,
     JobTerminatedError,
+    LogsError,
     PrepareCommandError,
     PrepareCommandTimeout,
     TestCommandError,
@@ -60,7 +62,7 @@ def _get_env_str(env: Dict[str, str]) -> str:
     return env_str
 
 
-class AnyscaleJobRunner(JobRunner):
+class AnyscaleJobRunner(CommandRunner):
     def __init__(
         self,
         cluster_manager: ClusterManager,
@@ -114,9 +116,14 @@ class AnyscaleJobRunner(JobRunner):
         self._artifact_path = artifact_path
         self._artifact_uploaded = artifact_path is not None
 
+    def _copy_script_to_working_dir(self, script_name):
+        script = os.path.join(os.path.dirname(__file__), f"_{script_name}")
+        shutil.copy(script, script_name)
+
     def prepare_remote_env(self):
         self._copy_script_to_working_dir("anyscale_job_wrapper.py")
-        super().prepare_remote_env()
+        self._copy_script_to_working_dir("wait_cluster.py")
+        self._copy_script_to_working_dir("prometheus_metrics.py")
 
     def run_prepare_command(
         self, command: str, env: Optional[Dict] = None, timeout: float = 3600.0
@@ -336,6 +343,12 @@ class AnyscaleJobRunner(JobRunner):
         )
 
         return time_taken
+
+    def get_last_logs_ex(self) -> Optional[str]:
+        try:
+            return self.job_manager.get_last_logs()
+        except Exception as e:
+            raise LogsError(f"Could not get last logs: {e}") from e
 
     def _fetch_json(self, path: str) -> Dict[str, Any]:
         try:
